requirements.txt
pandas==1.1.5
PyPDF2==2.10.6
camelot_py==0.10.1
openpyxl==3.0.10
python-docx==0.8.11
opencv-python==4.6.0.66